Just some refactoring of config parsing (#16)
Wanted to pull the parts out of EmsSinkConfig so we can re-use some of them in a new EmsStreamingSinkConfig
