+++ connector/src/main/scala/com/celonis/kafka/connect/ems/config/ProxyConfig.scala
@@ -3,13 +3,14 @@
  */
 package com.celonis.kafka.connect.ems.config
 
+import cats.syntax.either._
+import enumeratum._
 import org.asynchttpclient.AsyncHttpClient
 import org.asynchttpclient.DefaultAsyncHttpClient
 import org.asynchttpclient.DefaultAsyncHttpClientConfig
 import org.asynchttpclient.Realm
 import org.asynchttpclient.proxy.ProxyServer
 import org.asynchttpclient.proxy.{ ProxyType => AsyncProxyType }
-import enumeratum._
 
 sealed abstract class ProxyType(val proxyType: AsyncProxyType) extends EnumEntry
 
@@ -25,6 +26,40 @@ object ProxyType extends Enum[ProxyType] {
 
 sealed trait ProxyConfig {
   def createHttpClient(): AsyncHttpClient
+}
+
+object ProxyConfig {
+  import EmsSinkConfigConstants._
+
+  def extractProxyAuth(props: Map[String, _]): Either[String, Option[BasicAuthentication]] =
+    PropertiesHelper.getString(props, PROXY_AUTHENTICATION_KEY) match {
+      case Some("BASIC") => {
+          for {
+            user <- PropertiesHelper.getString(props, PROXY_AUTHBASIC_USERNAME_KEY)
+            pass <- PropertiesHelper.getPassword(props, PROXY_AUTHBASIC_PASSWORD_KEY)
+          } yield BasicAuthentication(user, pass)
+        }.asRight
+      case Some(other) => s"Proxy authentication type not currently supported: ($other). Supported values: BASIC".asLeft
+      case None        => None.asRight
+    }
+
+  def extractProxy(props: Map[String, _]): Either[String, ProxyConfig] = {
+    for {
+      host <- PropertiesHelper.getString(props, PROXY_HOST_KEY)
+      port <- PropertiesHelper.getInt(props, PROXY_PORT_KEY)
+      proxyType <- for {
+        proxyTypeStringOpt <- PropertiesHelper.getString(props, PROXY_TYPE_KEY)
+        proxyType           = ProxyType.withNameInsensitiveOption(proxyTypeStringOpt).getOrElse(ProxyType.Http)
+      } yield proxyType
+    } yield {
+      (host, port, proxyType)
+    }
+  }
+    .map {
+      case (host, port, proxyType) => ProxyConfig.extractProxyAuth(props)
+          .map(maybeAuth => ConfiguredProxyConfig(host, port, proxyType, maybeAuth))
+    }
+    .getOrElse(NoProxyConfig().asRight)
 }
 
 case class NoProxyConfig() extends ProxyConfig {
