+++ connector/src/test/scala/com/celonis/kafka/connect/ems/config/RetryTests.scala
@@ -12,13 +12,13 @@ import org.scalatest.matchers.should.Matchers
 
 class RetryTests extends AnyFunSuite with Matchers {
   test(s"return defaults if retry keys are missing") {
-    EmsSinkConfig.extractRetry(Map.empty) shouldBe Right(RetryConfig(NBR_OF_RETIRES_DEFAULT,
-                                                                     ERROR_RETRY_INTERVAL_DEFAULT,
+    RetryConfig.extractRetry(Map.empty) shouldBe Right(RetryConfig(NBR_OF_RETIRES_DEFAULT,
+                                                                   ERROR_RETRY_INTERVAL_DEFAULT,
     ))
   }
 
   test(s"return the retry config") {
-    EmsSinkConfig.extractRetry(Map(ERROR_RETRY_INTERVAL -> 1000, NBR_OF_RETRIES_KEY -> 10)) shouldBe Right(
+    RetryConfig.extractRetry(Map(ERROR_RETRY_INTERVAL -> 1000, NBR_OF_RETRIES_KEY -> 10)) shouldBe Right(
       RetryConfig(
         10,
         1000,
@@ -28,11 +28,11 @@ class RetryTests extends AnyFunSuite with Matchers {
 
   test(s"return an error if retry interval is smaller than 1s") {
     val message = s"Invalid [$ERROR_RETRY_INTERVAL]. Retry interval cannot be smaller than 1000 (1s)."
-    EmsSinkConfig.extractRetry(Map(ERROR_RETRY_INTERVAL -> 100, NBR_OF_RETRIES_KEY -> 10)) shouldBe Left(message)
+    RetryConfig.extractRetry(Map(ERROR_RETRY_INTERVAL -> 100, NBR_OF_RETRIES_KEY -> 10)) shouldBe Left(message)
   }
 
   test(s"return an error if retries is smaller than 1") {
     val message = s"Invalid [$NBR_OF_RETRIES_KEY]. Number of retries needs to be greater than 0."
-    EmsSinkConfig.extractRetry(Map(ERROR_RETRY_INTERVAL -> 1000, NBR_OF_RETRIES_KEY -> 0)) shouldBe Left(message)
+    RetryConfig.extractRetry(Map(ERROR_RETRY_INTERVAL -> 1000, NBR_OF_RETRIES_KEY -> 0)) shouldBe Left(message)
   }
 }
